requirements.txt
black==22.3.0
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
importlib-metadata==4.11.3
mypy-extensions==0.4.3
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.5.2
PyYAML==6.0
requests==2.27.1
tomli==2.0.1
typed-ast==1.5.3
typing-extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
